pixelcnn/requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
cycler==0.10.0
decorator==4.4.1
gast==0.3.2
grpcio==1.26.0
h5py==2.10.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.2
joblib==0.14.1
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.16.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.5
python-dateutil==2.8.1
pyzmq==18.1.1
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
sklearn==0.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
tornado==6.0.3
tqdm==4.41.0
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
